Specify tz with the date
This test was failing on my local machine because of TZ differences. Here's the breakdown.

1. System Date: `2022-05-01`. Jest interprets it as `2022-05-01 00:00:00 UTC`
2. 1hr before `2022-04-30 23:00:00 UTC`
3. Override value: `2022-04-30 22:00:00` which JavaScript interprets as PDT. Converted to UTC -> `2022-05-01 05:00:00 UTC`. That's definitely not before `2022-05-01 00:00:00 UTC`

Override Value: `2022-04-30 22:00:00` which is interpreted as `2022-04-30 22:00:00 PDT` -> `2022-05-01 05:00:00 UTC`. This is definitely not before `2022-04-30 16:00:00 PDT`

Some of the other tests might break for other TZ. We should probably mock the TZ in the tests to be more explicit. But I just want to make these tests pass my machine for now.