feat(flags): Add feature flag version info, evaluation reason, and id to `$feature_flag_called` event (#427)
* Suppress extraneous console output

It's a lot of noise when running tests.

* Fix example server to use passed in flag id

We pass in the flag id on the query string. Might as well use it!

* Add note about running tests

* Rename to indicate v3 tests

* Add type annotations and helper methods

* Annotate types in react-native lib

* Add emit event test and update bootstrap tests

If the bootstrap.featureFlagPayloads has a key
that's not in bootstrap.featureFlags, we assume
the caller wanted that to be a true flag.

* Normalize getDecide response

Handle v4 Response from decide

* Refactor access to persisted and bootstrapped properties

This changes how we load and save flags from persistence to support the new v4 format. However, we have to be careful to honor any flags already in persistence in the legacy format.

* Add some more helpers

* Implement `getFeatureFlagDetails`

* Reimplement getFeatureFlags and getFeatureFlagPayloads

Use getFeatureFlagDetails to reimplement it.

* Add extra info to $feature_flag_called event

The $feature_flag_called event now has additional information such as `feature_flag_id`, `feature_flag_version`, `feature_flag_reason`, and `feature_flag_requestId`

Fixes #427

* Update changelog and version

* Update README to link to supporting docs

* Revert change to react-native version

* Add an example that gets all flags and payloads

* Remove TODO statement

It's now a TODID.