docs: Add CHANGELOG notice about GZip compression (#564)
PostHog web now attempts to compress messages with GZip before sending them to our servers if available in your runtime. This shouldn't cause you any problems, but if someone's tests are attempting to assert on the data that's sent to our servers, they'll now need to decompress it first.

Let's note this on the BREAKING CHANGES section of both our web and node versions